Add ability to ingest hashdeep csv file into sqlite
This uses pandas for "simplicity", but because of hashdeep's failure to quote fields containing the
delimiter (the filename can contain ',' and is not quoted), I'm jumping through hoops here, and
maybe pandas becomes overkill.

Also would like to have the option to do this either in-memory or on-disk, but I need to figure out
the connection lifecycle in that optional world.